Moved extrapolation tests into job role test script, and test data/schema too
