+++ jobs/estimate_ind_cqc_filled_posts_by_job_role.py
@@ -13,7 +13,9 @@ from utils.estimate_filled_posts_by_job_role_utils import utils as JRutils
 from utils.estimate_filled_posts_by_job_role_utils.models.interpolation import (
     model_job_role_ratio_interpolation,
 )
-from utils.estimate_filled_posts.models.extrapolation import extrapolate_job_role_ratios
+from utils.estimate_filled_posts_by_job_role_utils.models.extrapolation import (
+    extrapolate_job_role_ratios,
+)
 from utils.ind_cqc_filled_posts_utils import utils as FPutils
 
 PartitionKeys = [Keys.year, Keys.month, Keys.day, Keys.import_date]
