+++ tests/test_file_schemas.py
@@ -2638,29 +2638,6 @@ class ModelExtrapolation:
         ]
     )
 
-    extrapolate_job_role_ratios_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), False),
-            StructField(IndCQC.unix_time, IntegerType(), False),
-            StructField(
-                IndCQC.ascwds_job_role_ratios_filtered,
-                MapType(StringType(), FloatType()),
-                True,
-            ),
-        ]
-    )
-
-    expected_extrapolate_job_role_ratios_schema = StructType(
-        [
-            *extrapolate_job_role_ratios_schema.fields,
-            StructField(
-                IndCQC.ascwds_job_role_ratios_extrapolated,
-                MapType(StringType(), FloatType()),
-                True,
-            ),
-        ]
-    )
-
 
 @dataclass
 class ModelInterpolation:
@@ -6214,6 +6191,28 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsSchemas:
             *interpolate_job_role_ratios_schema,
             StructField(
                 IndCQC.ascwds_job_role_ratios_interpolated,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+
+    extrapolate_job_role_ratios_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(
+                IndCQC.ascwds_job_role_ratios_filtered,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+    expected_extrapolate_job_role_ratios_schema = StructType(
+        [
+            *extrapolate_job_role_ratios_schema.fields,
+            StructField(
+                IndCQC.ascwds_job_role_ratios_extrapolated,
                 MapType(StringType(), FloatType()),
                 True,
             ),
