+++ tests/unit/test_extrapolation.py
@@ -317,50 +317,3 @@ class CombineExtrapolationTests(ModelExtrapolationTests):
                 self.expected_data[i][IndCqc.extrapolation_model],
                 f"Returned value in row {i} does not match expected",
             )
-
-
-class JobRoleRatiosExtrapolationTests(ModelExtrapolationTests):
-    def setUp(self):
-        super().setUp()
-        self.input_df = self.spark.createDataFrame(
-            Data.job_role_ratios_extrapolation_rows,
-            Schemas.extrapolate_job_role_ratios_schema,
-        )
-        self.expected_df = self.spark.createDataFrame(
-            Data.expected_job_role_ratios_extrapolation_rows,
-            Schemas.expected_extrapolate_job_role_ratios_schema,
-        )
-        self.returned_df = job.extrapolate_job_role_ratios(self.input_df)
-
-        self.returned_data = (
-            self.returned_df.select(
-                IndCqc.location_id,
-                IndCqc.unix_time,
-                IndCqc.ascwds_job_role_ratios_filtered,
-                IndCqc.ascwds_job_role_ratios_extrapolated,
-            )
-            .sort(IndCqc.location_id, IndCqc.unix_time)
-            .collect()
-        )
-
-        self.expected_data = (
-            self.expected_df.select(
-                IndCqc.location_id,
-                IndCqc.unix_time,
-                IndCqc.ascwds_job_role_ratios_filtered,
-                IndCqc.ascwds_job_role_ratios_extrapolated,
-            )
-            .sort(IndCqc.location_id, IndCqc.unix_time)
-            .collect()
-        )
-
-    def test_ratios_extrapolation_returns_same_row_count(self):
-        self.assertEqual(self.input_df.count(), self.returned_df.count())
-
-    def test_ratios_extrapolation_returns_expected_column(self):
-        self.assertIn(
-            IndCqc.ascwds_job_role_ratios_extrapolated, self.returned_df.columns
-        )
-
-    def test_ratios_extrapolation_values_match_expected(self):
-        self.assertEqual(self.returned_data, self.expected_data)
